test(agent): extract shared ClaudeAcpAgent test harness
The idle-drain, streamed-text, and slash-command tests each carried their own
copies of installFakeSession and the message-builder helpers. Extract them into
a shared harness (packages/agent/src/test/helpers/claude-agent.ts) and import
from all three, unifying the drifted installFakeSession into one definition
(returns {query, input}, optional knownSlashCommands, superset of inert fields).

makeAgent and the vi.mock blocks stay per-file — they depend on the
dynamically-imported (post-vi.mock) agent class, so moving them risks mock
hoisting surprises. File-specific helpers (streamed-text's messageStart/
textDelta/enableSignedCommitServer, slash-command's findUnsupportedChunkText)
stay put. No behavior change.

Generated-By: PostHog Code
Task-Id: 39548658-f313-445b-bd1c-d07eaf9281cf
