+++ packages/agent/src/adapters/claude/claude-agent.idle-drain.test.ts
@@ -1,11 +1,19 @@
 import type { AgentSideConnection } from "@agentclientprotocol/sdk";
-import type {
-  SDKMessage,
-  SDKUserMessage,
-} from "@anthropic-ai/claude-agent-sdk";
+import type { SDKUserMessage } from "@anthropic-ai/claude-agent-sdk";
 import { beforeEach, describe, expect, it, vi } from "vitest";
-import { createMockQuery, type MockQuery } from "../../test/mocks/claude-sdk";
-import { Pushable } from "../../utils/streams";
+import {
+  assistantMessage,
+  type ClientMocks,
+  echoUserMessage,
+  installFakeSession,
+  makeClientMocks,
+  messageChunkTexts,
+  resultSuccess,
+  send,
+  tick,
+} from "../../test/helpers/claude-agent";
+import type { MockQuery } from "../../test/mocks/claude-sdk";
+import type { Pushable } from "../../utils/streams";
 
 vi.mock("@anthropic-ai/claude-agent-sdk", () => ({
   query: vi.fn(),
@@ -25,124 +33,10 @@ vi.mock("./mcp/tool-metadata", () => ({
 const { ClaudeAcpAgent } = await import("./claude-agent");
 type Agent = InstanceType<typeof ClaudeAcpAgent>;
 
-interface ClientMocks {
-  sessionUpdate: ReturnType<typeof vi.fn>;
-  extNotification: ReturnType<typeof vi.fn>;
-}
-
 function makeAgent(): { agent: Agent; client: ClientMocks } {
-  const client: ClientMocks = {
-    sessionUpdate: vi.fn().mockResolvedValue(undefined),
-    extNotification: vi.fn().mockResolvedValue(undefined),
-  };
+  const client = makeClientMocks();
   const agent = new ClaudeAcpAgent(client as unknown as AgentSideConnection);
   return { agent, client };
-}
-
-function installFakeSession(
-  agent: Agent,
-  sessionId: string,
-): { query: MockQuery; input: Pushable<SDKUserMessage> } {
-  const query = createMockQuery();
-  const input = new Pushable<SDKUserMessage>();
-  const abortController = new AbortController();
-
-  const session = {
-    query,
-    queryOptions: { sessionId, cwd: "/tmp/repo", abortController },
-    buildInProcessMcpServers: () => ({}),
-    localToolsServerNames: [] as string[],
-    input,
-    cancelled: false,
-    interruptReason: undefined,
-    settingsManager: { dispose: vi.fn(), getRepoRoot: () => "/tmp/repo" },
-    permissionMode: "default" as const,
-    abortController,
-    accumulatedUsage: {
-      inputTokens: 0,
-      outputTokens: 0,
-      cachedReadTokens: 0,
-      cachedWriteTokens: 0,
-    },
-    sessionResources: new Set(),
-    configOptions: [],
-    promptRunning: false,
-    pendingMessages: new Map(),
-    nextPendingOrder: 0,
-    cwd: "/tmp/repo",
-    notificationHistory: [] as unknown[],
-    taskRunId: "run-1",
-    lastContextWindowSize: 200_000,
-    modelId: "claude-sonnet-4-6",
-    taskState: new Map(),
-  };
-
-  (agent as unknown as { session: typeof session }).session = session;
-  (agent as unknown as { sessionId: string }).sessionId = sessionId;
-
-  return { query, input };
-}
-
-function tick(): Promise<void> {
-  return new Promise((resolve) => setImmediate(resolve));
-}
-
-async function send(query: MockQuery, message: unknown): Promise<void> {
-  query._mockHelpers.sendMessage(message as SDKMessage);
-  await tick();
-}
-
-// Replays the prompt's own user message back through the query so
-// `promptReplayed` flips and the terminal `result` is not skipped.
-async function echoUserMessage(
-  query: MockQuery,
-  input: Pushable<SDKUserMessage>,
-): Promise<void> {
-  const { value: pushed } = await input[Symbol.asyncIterator]().next();
-  await send(query, pushed);
-}
-
-function assistantMessage(sessionId: string, apiId: string, text: string) {
-  return {
-    type: "assistant",
-    parent_tool_use_id: null,
-    session_id: sessionId,
-    uuid: `assistant-${apiId}`,
-    message: {
-      id: apiId,
-      role: "assistant",
-      content: [{ type: "text", text }],
-    },
-  };
-}
-
-function resultSuccess(sessionId: string, uuid = "result-1") {
-  return {
-    type: "result",
-    subtype: "success",
-    session_id: sessionId,
-    uuid,
-    result: "",
-    is_error: false,
-    usage: {},
-    modelUsage: {},
-  };
-}
-
-function messageChunkTexts(
-  calls: ClientMocks["sessionUpdate"]["mock"]["calls"],
-): string[] {
-  return calls
-    .map(
-      ([call]) =>
-        (
-          call as {
-            update?: { sessionUpdate?: string; content?: { text?: string } };
-          }
-        ).update,
-    )
-    .filter((update) => update?.sessionUpdate === "agent_message_chunk")
-    .map((update) => update?.content?.text ?? "");
 }
 
 // Runs one complete turn: prompt -> echo -> assistant text -> result.
